+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/stub/type/ProtobufMessageStubType.kt
@@ -1,9 +1,11 @@
 package io.kanro.idea.plugin.protobuf.lang.psi.stub.type
 
+import com.intellij.psi.stubs.IndexSink
 import com.intellij.psi.stubs.StubElement
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufMessageDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.impl.ProtobufMessageDefinitionImpl
 import io.kanro.idea.plugin.protobuf.lang.psi.stub.impl.ProtobufMessageStub
+import io.kanro.idea.plugin.protobuf.lang.psi.stub.index.ResourceTypeIndex
 
 object ProtobufMessageStubType : ProtobufStubTypeBase<ProtobufMessageStub, ProtobufMessageDefinition>(
     "MESSAGE_DEFINITION"
@@ -18,5 +20,12 @@ object ProtobufMessageStubType : ProtobufStubTypeBase<ProtobufMessageStub, Proto
 
     override fun createPsi(stub: ProtobufMessageStub): ProtobufMessageDefinition {
         return ProtobufMessageDefinitionImpl(stub, stub.stubType)
+    }
+
+    override fun indexStub(stub: ProtobufMessageStub, sink: IndexSink) {
+        super.indexStub(stub, sink)
+        stub.resourceType()?.let {
+            sink.occurrence(ResourceTypeIndex.key, it)
+        }
     }
 }